+++ fj-daogen-base/src/main/java/org/fugerit/java/daogen/base/gen/DaogenBasicGenerator.java
@@ -471,7 +471,14 @@ public abstract class DaogenBasicGenerator extends SimpleJavaGenerator implement
 			}
 		}
 	}
-	
+
+	protected void generateSerial( boolean condition ) throws IOException {
+		if ( condition ) {
+			this.addSerialVerUID();
+			this.writeSerialHelpers();
+		}
+	}
+
 	protected void writeSerialHelpers() {
 		this.getWriter().println( TAB+"private void writeObject(java.io.ObjectOutputStream out) throws java.io.IOException {" );
 		this.getWriter().println( TAB_2+"// this class is conditionally serializable, depending on contained object" );