+++ fj-daogen-sample/src/main/java/org/fugerit/java/daogen/sample/impl/struct/ObjUpload.java
@@ -118,7 +118,7 @@ public class ObjUpload extends WrapperUpload implements SQLData, StructMapper {
 	@Override
 	public void writeSQL(SQLOutput stream) throws SQLException {
 		if ( !this.checkLobs() ) {
-			throwUnsupported( "To use writeSQL() you must invoke setupLobs() for  "+this.getSQLTypeName() );
+			org.fugerit.java.core.db.daogen.BasicHelper.throwUnsupported( "To use writeSQL() you must invoke setupLobs() for  "+this.getSQLTypeName() );
 		}
 		this.areLobsSet = false;	// clob and blob will be used only once
 		stream.writeBigDecimal( this.getId() );